requirements.txt
cryptography>=3.4.7
python-dotenv>=0.19.2
pyfiglet>=0.8.post1
pyperclip>=1.8.2
